# api_gateway/api/docs/swagger.yaml
definitions:
...
  models.AddressRequest:
    properties:
      house_number:
        type: integer
      street:
        type: string
    type: object
...
  models.CustomerRegister:
    properties:
      addresses:
        items:
          $ref: '#/definitions/models.AddressRequest'
        type: array
      bio:
        description: '`json:"username"`'
        type: string
      email:
        description: '`json:"email"`'
        type: string
      firstName:
        description: '`json:"first_name"`'
        type: string
      lastName:
        description: '`json:"last_name"`'
        type: string
      password:
        description: '`json:"password"`'
        type: string
    type: object
  models.Error:
    properties:
      code:
        type: integer
      description:
        type: string
      error: {}
    type: object
...
  review.ReviewRequest:
    properties:
      customer_id:
        type: integer
      description:
        type: string
      name:
        type: string
      post_id:
        type: integer
      review:
        type: integer
    type: object
info:
  contact:
    email: rahimzanovmuhammadumar@gmail.com
    name: Muhammad Umar
    url: https://t.me/muhammad_ummar
  description: This is exam server api server
  termsOfService: 2 term exam
  title: exam api
  version: "1.0"
paths:
...
  /v1/register:
    post:
      consumes:
      - application/json
      parameters:
      - description: user data
        in: body
        name: userData
        required: true
        schema:
          $ref: '#/definitions/models.CustomerRegister'
      produces:
      - application/json
      responses:
        "200":
          description: Message sended to your email succesfully
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Error'
      summary: Register for authentication
      tags:
      - Auth
  /v1/register/{code}/{email}:
    get:
      consumes:
      - application/json
      parameters:
      - description: code
        in: path
        name: code
        required: true
        type: integer
      - description: email
        in: path
        name: email
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: json
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Error'
      summary: Verify for authentication
      tags:
      - Auth
  /v1/review:
    post:
      consumes:
      - application/json
      description: this api creates new review
      parameters:
      - description: Review
        in: body
        name: review
        required: true
        schema:
          $ref: '#/definitions/review.ReviewRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            type: json
      summary: create review api
      tags:
      - review
  /v1/review/{id}:
    delete:
      consumes:
      - application/json
      description: this api deletes review from database
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: integer
      responses: {}
      security:
      - BearerAuth: []
      summary: Delete review api
      tags:
      - review
    get:
      consumes:
      - application/json
      description: this api gets review from database
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: integer
      responses: {}
      security:
      - BearerAuth: []
      summary: Get review api
      tags:
      - review
  /v1/review_by_custID/{id}:
    delete:
      consumes:
      - application/json
      description: this api deletes review by customer
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: json
      security:
      - BearerAuth: []
      summary: delete review by cust api
      tags:
      - review
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"